backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
psycopg2-binary>=2.9
django-cors-headers>=4.3
python-dotenv>=1.0
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
djangorestframework-simplejwt>=5.3
djangorestframework-simplejwt==5.3.1
djangorestframework
djangorestframework-simplejwt